[FFI] Fix system library symbol lookup
This PR fixes a bug in the current system library symbol lookup.

Prior to this PR, the lookup key always concatenates `symbol_prefix_`
and `name`. However, the `name` may or may not have already contained
the symbol prefix. And when it does contain the symbol prefix, the
concantenated string will have duplicate symbol prefix, which leads
to the lookup failure, because in the compiled library the symbol
prefix only appears once.

This PR changes the lookup behavior to check both cases.